refactor to fucntional components
