+++ crates/mipsy_web/src/components/mod.rs
@@ -2,3 +2,5 @@ pub mod navbar;
 pub mod pagebackground;
 pub mod modal;
 pub mod outputarea;
+pub mod sourcecode;
+pub mod decompiled;
