+++ crates/mipsy_web/src/pages/main/mod.rs
@@ -1,0 +1,3 @@
+pub mod app;
+pub mod update;
+pub mod state;
