+++ crates/mipsy_web/src/pages/mod.rs
@@ -1,0 +1,1 @@
+pub mod main;
